+++ src/transport/transport.c
@@ -1,6 +1,7 @@
 /* SPDX-License-Identifier: Apache-2.0 OR GPL-2.0-or-later */
 #include "compiler.h"
 #include "transport.h"
+#include "environ/time.h"
 
 #include <errno.h>
 #include <libpldm/base.h>
@@ -114,13 +115,15 @@ static long timeval_to_msec(const struct timeval *tv)
 /* If calculations on `tv` don't overflow then operations on derived
  * intervals can't either.
  */
-static bool timeval_is_valid(const struct timeval *tv)
+static bool timeval_validate_for_msec(const struct timeval *tv)
 {
+	/* Must be a normalised, positive interval */
 	if (tv->tv_sec < 0 || tv->tv_usec < 0 || tv->tv_usec >= 1000000) {
 		return false;
 	}
 
-	if (tv->tv_sec > (LONG_MAX - tv->tv_usec / 1000) / 1000) {
+	/* Components must safely convert to msec */
+	if (tv->tv_sec > ((LONG_MAX / 1000) - 1000)) {
 		return false;
 	}
 
@@ -132,7 +135,7 @@ static int clock_gettimeval(clockid_t clockid, struct timeval *tv)
 	struct timespec now;
 	int rc;
 
-	rc = clock_gettime(clockid, &now);
+	rc = libpldm_clock_gettime(clockid, &now);
 	if (rc < 0) {
 		return rc;
 	}
@@ -201,14 +204,15 @@ pldm_transport_send_recv_msg(struct pldm_transport *transport, pldm_tid_t tid,
 	}
 
 	timeradd(&now, &max_response_interval, &end);
-	if (!timeval_is_valid(&end)) {
-		return PLDM_REQUESTER_POLL_FAIL;
-	}
 
 	while (timercmp(&now, &end, <)) {
 		pldm_tid_t src_tid;
 
 		timersub(&end, &now, &remaining);
+
+		if (!timeval_validate_for_msec(&remaining)) {
+			return PLDM_REQUESTER_POLL_FAIL;
+		}
 
 		/* 0 <= `timeval_to_msec()` <= 4800, and 4800 < INT_MAX */
 		ret = pldm_transport_poll(transport,